+++ specification/ilm/_types/Policy.ts
@@ -17,9 +17,10 @@
  * under the License.
  */
 
-import { Phases } from './Phases'
+import { Name } from '@_types/common'
+import { Phases } from './Phase'
 
 export class Policy {
   phases: Phases
-  name?: string
+  name?: Name
 }
