+++ specification/ilm/delete_lifecycle/DeleteLifecycleRequest.ts
@@ -30,6 +30,4 @@ export interface Request extends RequestBase {
     policy?: Name
     policy_id: Id
   }
-  query_parameters?: {}
-  body?: {}
 }
