+++ specification/ilm/get_lifecycle/GetLifecycleRequest.ts
@@ -30,6 +30,4 @@ export interface Request extends RequestBase {
     policy?: Name
     policy_id?: Id
   }
-  query_parameters?: {}
-  body?: {}
 }
